Moved `ReorderPage` to `pages` for consistency
