+++ lib/pages/reorder_properties_page.dart
@@ -1,6 +1,6 @@
 import 'package:converterpro/models/order.dart';
 import 'package:converterpro/data/default_order.dart';
-import 'package:converterpro/utils/reorder_page.dart';
+import 'package:converterpro/pages/reorder_page.dart';
 import 'package:converterpro/pages/splash_screen.dart';
 import 'package:converterpro/data/property_unit_maps.dart';
 import 'package:converterpro/utils/utils.dart';
